+++ image.cpp
@@ -48,14 +48,14 @@ void Image::CreateTexture()
 Image Image::Blend(Image& a, Image& b, float pct)
 {
 	Image result(a.m_Width, a.m_Height, a.m_Channels);
-	for (int y = 0; y < a.m_Height; y++) {
-		for (int x = 0; x < a.m_Width; x++) {
+	for (size_t y = 0; y < a.m_Height; y++) {
+		for (size_t x = 0; x < a.m_Width; x++) {
 			glm::ivec3 pixelA = a(x, y);
 			glm::ivec3 pixelB = b(x, y);
 			glm::ivec3 blendedPixel;
-			blendedPixel.r = (1.0f - pct) * pixelA.r + pct * pixelB.r;
-			blendedPixel.g = (1.0f - pct) * pixelA.g + pct * pixelB.g;
-			blendedPixel.b = (1.0f - pct) * pixelA.b + pct * pixelB.b;
+			blendedPixel.r = int((1.0f - pct) * pixelA.r + pct * pixelB.r);
+			blendedPixel.g = int((1.0f - pct) * pixelA.g + pct * pixelB.g);
+			blendedPixel.b = int((1.0f - pct) * pixelA.b + pct * pixelB.b);
 			unsigned char* resultPixel = result.m_Data + (result.m_Channels * (y * result.m_Width + x));
 			resultPixel[0] = blendedPixel.r;
 			resultPixel[1] = blendedPixel.g;
